Move backend encoder
